+++ kwimage/structs/mask.py
@@ -161,6 +161,13 @@ class _MaskConversionMixin(object):
             format (str):
                 the string code for the format you want to transform into.
 
+            copy (bool, default=True):
+                if True, we always return copy of the data.
+                if False, we try not to return a copy unless necessary.
+
+        Returns:
+            Mask: The Mask object with a new backend format
+
         Example:
             >>> # xdoctest: +REQUIRES(--mask)
             >>> from kwimage.structs.mask import MaskFormat  # NOQA
@@ -190,6 +197,16 @@ class _MaskConversionMixin(object):
     @_register_convertor(MaskFormat.BYTES_RLE)
     def to_bytes_rle(self, copy=False):
         """
+        Converts the mask format to a bytes-based run-length encoding.
+
+        Args:
+            copy (bool, default=True):
+                if True, we always return copy of the data.
+                if False, we try not to return a copy unless necessary.
+
+        Returns:
+            Mask: The Mask object with a new backend format
+
         Example:
             >>> # xdoctest: +REQUIRES(--mask)
             >>> from kwimage.structs.mask import MaskFormat  # NOQA
@@ -238,6 +255,17 @@ class _MaskConversionMixin(object):
 
     @_register_convertor(MaskFormat.ARRAY_RLE)
     def to_array_rle(self, copy=False):
+        """
+        Converts the mask format to an array-based run-length encoding.
+
+        Args:
+            copy (bool, default=True):
+                if True, we always return copy of the data.
+                if False, we try not to return a copy unless necessary.
+
+        Returns:
+            Mask: the underlying RLE data will be in F-contiguous order.
+        """
         if self.format == MaskFormat.ARRAY_RLE:
             return self.copy() if copy else self
         elif self.format == MaskFormat.BYTES_RLE:
@@ -333,6 +361,17 @@ class _MaskConversionMixin(object):
 
     @_register_convertor(MaskFormat.C_MASK)
     def to_c_mask(self, copy=False):
+        """
+        Convert the mask format to a dense mask array in rowwise (C) order
+
+        Args:
+            copy (bool, default=True):
+                if True, we always return copy of the data.
+                if False, we try not to return a copy unless necessary.
+
+        Returns:
+            Mask: The Mask object with a new backend format
+        """
         if self.format == MaskFormat.C_MASK:
             return self.copy() if copy else self
         elif self.format == MaskFormat.F_MASK:
@@ -347,6 +386,9 @@ class _MaskConversionMixin(object):
     def numpy(self):
         """
         Ensure mask is in numpy format (if possible)
+
+        Returns:
+            Mask: The Mask object with a new backend format
         """
         data = self.data
         if self.format in {MaskFormat.C_MASK, MaskFormat.F_MASK}:
@@ -358,6 +400,9 @@ class _MaskConversionMixin(object):
     def tensor(self, device=ub.NoParam):
         """
         Ensure mask is in tensor format (if possible)
+
+        Returns:
+            Mask: The Mask object with a new backend format
         """
         data = self.data
         if self.format in {MaskFormat.C_MASK, MaskFormat.F_MASK}:
@@ -377,13 +422,16 @@ class _MaskConstructorMixin(object):
     @classmethod
     def from_polygons(Mask, polygons, dims):
         """
-        DEPRICATE: use kwimage.Polygon.to_mask?
+        DEPRICATE: use kwimage.Polygon.to_mask? or kwimage.Mask.coerce?
 
         Args:
             polygons (ndarray | List[ndarray]): one or more polygons that
                 will be joined together. The ndarray may either be an
                 Nx2 or a flat c-contiguous array or xy points.
             dims (Tuple): height / width of the source image
+
+        Returns:
+            Mask: the new Mask object
 
         Example:
             >>> # xdoctest: +REQUIRES(--mask)
@@ -475,6 +523,21 @@ class _MaskTransformMixin(object):
     @profile
     def scale(self, factor, output_dims=None, inplace=False):
         """
+        Perform a scale operation on the mask.
+
+        Args:
+            factor (float | Tuple[float, float]): the xy scale factor
+
+            input_dims (Tuple[int, int]): unused
+
+            output_dims (Tuple[int, int]): shape of the returned mask
+
+        Returns:
+            Mask: the transformed Mask object
+
+        Notes:
+            * This function has not been optimized and may be inefficient
+
         Example:
             >>> # xdoctest: +REQUIRES(module:torch)
             >>> self = Mask.random()
@@ -497,6 +560,21 @@ class _MaskTransformMixin(object):
     # @profile
     def warp(self, transform, input_dims=None, output_dims=None, inplace=False):
         """
+        Perform a matrix warp (e.g. affine or projective) on the underlying
+        mask data.
+
+        Args:
+            transform (ndarray): the transform matrix
+
+            input_dims (Tuple[int, int]): unused
+
+            output_dims (Tuple[int, int]): shape of the returned mask
+
+        Returns:
+            Mask: the transformed Mask object
+
+        Notes:
+            * This function has not been optimized and may be inefficient
 
         Example:
             >>> # xdoctest: +REQUIRES(module:torch)
@@ -522,6 +600,8 @@ class _MaskTransformMixin(object):
         if transform is None:
             new = self if inplace else Mask(self.data.copy(), self.format)
             return new
+        elif isinstance(transform, kwimage.Affine):
+            transform = transform.matrix
 
         c_mask = self.to_c_mask(copy=False).data
         t_mask = torch.Tensor(c_mask)
@@ -545,6 +625,9 @@ class _MaskTransformMixin(object):
                 If unspecified the parent shape is used.
 
             inplace (bool): for api compatability, currently ignored
+
+        Returns:
+            Mask: the transformed Mask object
 
         Example:
             >>> self = Mask.random(shape=(8, 8), rng=0)
@@ -770,6 +853,16 @@ class Mask(ub.NiceRepr, _MaskConversionMixin, _MaskConstructorMixin,
     @classmethod
     def random(Mask, rng=None, shape=(32, 32)):
         """
+        Create a random binary mask object
+
+        Args:
+            rng (int | RandomState | None): the random seed
+
+            shape (Tuple[int, int]): the height / width of the returned mask
+
+        Returns:
+            Mask: the random mask
+
         Example:
             >>> import kwimage
             >>> mask = kwimage.Mask.random()
@@ -793,6 +886,9 @@ class Mask(ub.NiceRepr, _MaskConversionMixin, _MaskConstructorMixin,
     def demo(cls):
         """
         Demo mask with holes and disjoint shapes
+
+        Returns:
+            Mask: the demo mask
         """
         text = ub.codeblock(
             '''
@@ -830,6 +926,9 @@ class Mask(ub.NiceRepr, _MaskConversionMixin, _MaskConstructorMixin,
         """
         Performs a deep copy of the mask data
 
+        Returns:
+            Mask: the copied mask
+
         Example:
             >>> self = Mask.random(shape=(8, 8), rng=0)
             >>> other = self.copy()
@@ -840,6 +939,12 @@ class Mask(ub.NiceRepr, _MaskConversionMixin, _MaskConstructorMixin,
     def union(self, *others):
         """
         This can be used as a staticmethod or an instancemethod
+
+        Args:
+            *others: multiple input masks to union
+
+        Returns:
+            Mask: the unioned mask
 
         Example:
             >>> # xdoc: +REQUIRES(--mask)
@@ -909,12 +1014,16 @@ class Mask(ub.NiceRepr, _MaskConversionMixin, _MaskConstructorMixin,
                     new_rle['size'] = list(map(int, new_rle['size']))  # python2 fix
                 new = cls(new_rle, MaskFormat.BYTES_RLE)
         return new
-        # rle_datas = [item.to_bytes_rle().data for item in items]
-        # return cls(cython_mask.merge(rle_datas, intersect=0), MaskFormat.BYTES_RLE)
 
     def intersection(self, *others):
         """
         This can be used as a staticmethod or an instancemethod
+
+        Args:
+            *others: multiple input masks to intersect
+
+        Returns:
+            Mask: the intersection of the masks
 
         Example:
             >>> n = 3
@@ -1294,6 +1403,7 @@ class Mask(ub.NiceRepr, _MaskConversionMixin, _MaskConstructorMixin,
             >>> polys.draw(border=True, linewidth=5, alpha=0.5, radius=0.2)
         """
         import cv2
+        from kwimage.structs.polygon import Polygon, MultiPolygon
         p = 2
         # It should be faster to only exact the patch of non-zero values
         x, y, w, h = self.get_xywh().astype(int).tolist()
@@ -1340,11 +1450,9 @@ class Mask(ub.NiceRepr, _MaskConversionMixin, _MaskConstructorMixin,
                     #     raise Exception
                     polys[i]['exterior'] = coords
 
-            from kwimage.structs.polygon import Polygon, MultiPolygon
             poly_list = [Polygon(**data) for data in polys.values()]
             multi_poly = MultiPolygon(poly_list)
         else:
-            from kwimage.structs.polygon import Polygon, MultiPolygon
             multi_poly = MultiPolygon([])
         return multi_poly
 
@@ -1412,7 +1520,7 @@ class Mask(ub.NiceRepr, _MaskConversionMixin, _MaskConstructorMixin,
                 height / width of the source image
 
         Returns:
-            Mask
+            Mask: the constructed mask object
 
         Example:
             >>> # xdoc: +REQUIRES(--mask)
@@ -1439,14 +1547,56 @@ class Mask(ub.NiceRepr, _MaskConversionMixin, _MaskConstructorMixin,
 
     def to_coco(self, style='orig'):
         """
-        Convert the Mask into a COCO json representation.
+        Convert the Mask to a COCO json representation based on the current
+        format.
+
+        A COCO mask is formatted as a run-length-encoding (RLE), of which there
+        are two variants: (1) a array RLE, which is slightly more readable and
+        extensible, and (2) a bytes RLE, which is slightly more concise. The
+        returned format will depend on the current format of the Mask object.
+        If it is in "bytes_rle" format, it will be returned in that format,
+        otherwise it will be converted to the "array_rle" format and returned
+        as such.
 
         Args:
-            style (str): Does nothing for this particular method. Always
-            converts based on the current format.
+            style (str): Does nothing for this particular method, exists for
+                API compatibility and if alternate encoding styles are
+                implemented in the future.
 
         Returns:
-            dict: either a bytes-rle or array-rle encoding
+            dict: either a bytes-rle or array-rle encoding, depending
+                on the current mask format. The keys in this dictionary
+                are as follows:
+
+                counts (List[int] | str): the array or bytes rle encoding
+
+                size (Tuple[int]): the height and width of the encoded mask
+                    *see note*.
+
+                shape (Tuple[int]): only present in array-rle mode. This
+                    is also the height/width of the underlying encoded array.
+                    This exists for semantic consistency with other kwimage
+                    conventions, and is not part of the original coco spec.
+
+                order (str): only present in array-rle mode.
+                    Either C or F, indicating if counts is aranged in row-major
+                    or column-major order. For COCO-compatibility this is
+                    always returned in F (column-major) order.
+
+                binary (bool): only present in array-rle mode.
+                    For COCO-compatibility this is always returned as False,
+                    indicating the mask only contains binary 0 or 1 values.
+
+        Note:
+            The output dictionary will contain a key named "size", this is the
+            only location in kwimage where "size" refers to a tuple in
+            (height/width) order, in order to be backwards compatible with the
+            original coco spec. In all other locations in kwimage a "size" will
+            refer to a (width/height) ordered tuple.
+
+        SeeAlso:
+            :func: kwimage.im_runlen.encode_run_length - backend function that
+                does array-style run length encoding.
 
         Example:
             >>> # xdoc: +REQUIRES(--mask)
@@ -1456,6 +1606,17 @@ class Mask(ub.NiceRepr, _MaskConversionMixin, _MaskConstructorMixin,
             >>> coco_data2 = self.toformat('bytes_rle').to_coco()
             >>> print('coco_data1 = {}'.format(ub.repr2(coco_data1, nl=1)))
             >>> print('coco_data2 = {}'.format(ub.repr2(coco_data2, nl=1)))
+            coco_data1 = {
+                'binary': True,
+                'counts': [47, 5, 3, 1, 14, ... 1, 4, 19, 141],
+                'order': 'F',
+                'shape': (23, 32),
+                'size': (23, 32),
+            }
+            coco_data2 = {
+                'counts': '_153L;4EL...ON3060L0N060L0Nb0Y4',
+                'size': [23, 32],
+            }
         """
         use_bytes = (self.format == MaskFormat.BYTES_RLE)
         if use_bytes:
@@ -1487,6 +1648,9 @@ class MaskList(_generic.ObjectList):
     def to_polygon_list(self):
         """
         Converts all mask objects to multi-polygon objects
+
+        Returns:
+            kwimage.PolygonList
         """
         import kwimage
         new = kwimage.PolygonList([
@@ -1498,6 +1662,9 @@ class MaskList(_generic.ObjectList):
     def to_segmentation_list(self):
         """
         Converts all items to segmentation objects
+
+        Returns:
+            kwimage.SegmentationList
         """
         import kwimage
         new = kwimage.SegmentationList([
@@ -1509,6 +1676,9 @@ class MaskList(_generic.ObjectList):
     def to_mask_list(self):
         """
         returns this object
+
+        Returns:
+            kwimage.MaskList
         """
         return self